src/main/java/com/knkweb/yarnshop/security/config/MyUserDetailService.java
package com.knkweb.yarnshop.security.config;

import com.knkweb.yarnshop.repositories.AuthorityRepository;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

@Service
public class MyUserDetailService implements UserDetailsService {

    private final AuthorityRepository authorityRepository;

    public MyUserDetailService(AuthorityRepository authorityRepository) {
        this.authorityRepository = authorityRepository;
    }


    @Override
    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
        return new myUserDetails(username);
    }
}
